Update 03.txt
Be watchful, and strengthen the things which remain, that are ready to die: for I have not found thy works perfect before God. ( kjv ) 

Thou hast a few names even in Sardis which have not defiled their garments; and they shall walk with me in white: for they are worthy. ( kjv )